Cast enum to uint32_t

+++ dtcInterfaceLib/DTC_Registers.cpp
@@ -4813,7 +4813,7 @@ void DTCLib::DTC_Registers::WriteRegister_(uint32_t data, const DTC_Register& ad
 	} while (retry > 0 && errorCode != 0);
 	if (errorCode != 0)
 	{
-		TLOG(TLVL_ERROR) << "Error writing register 0x" << std::hex << address << " " << errorCode;
+		TLOG(TLVL_ERROR) << "Error writing register 0x" << std::hex << static_cast<uint32_t>(address) << " " << errorCode;
 		throw DTC_IOErrorException(errorCode);
 	}
 }
@@ -4830,7 +4830,7 @@ uint32_t DTCLib::DTC_Registers::ReadRegister_(const DTC_Register& address)
 	} while (retry > 0 && errorCode != 0);
 	if (errorCode != 0)
 	{
-		TLOG(TLVL_ERROR) << "Error reading register 0x" << std::hex << address << " " << errorCode;
+		TLOG(TLVL_ERROR) << "Error reading register 0x" << std::hex << static_cast<uint32_t>(address) << " " << errorCode;
 		throw DTC_IOErrorException(errorCode);
 	}
